better allow.nfsd processing
we add the option unconditionally, but this option may not always
be present in the kernel - we will handle this.

While here - improved the output when starting the container,
so that the user can see the reason why the container does not start.

Pointed out: Mike Upchurch via Telegram, Thanks!